Split classes for easier use while constructing grouping method.
