+++ FETP/FETP/FETP.csproj
@@ -43,12 +43,16 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Block.cs" />
+    <Compile Include="Class.cs" />
+    <Compile Include="CourseInformation.cs" />
     <Compile Include="GA_Constraints.cs" />
     <Compile Include="FETP_Controller.cs" />
     <Compile Include="GA_Chromosome.cs" />
     <Compile Include="GA_Controller.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Schedule.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
